Do json serialization
Send json values between main window and xi thread, not just strings.
Also some various cleanups (including upgrading serde to 1.0).

Note that the public interface of xi-core-lib is still in terms of
strings, but since json_serde::Value is Send, it would make lots of
sense to skip the serialization and just send those. But that's
strictly a performance question.