Draw a cursor
Changes from just draw_text to using a TextLayout object, and does hit
testing on that, to decide where to draw the cursor.

Still very hacky (doesn't deal with utf-8 <-> utf-16 conversions, and
only a tiny fragment of the protocol implemented, but it's starting to
look a bit line an editor.