Start to factor out app shell
This commit creates a separate crate to abstract out a generic app
shell for winapi, taking care of window creation, lifecycle, and
related topics.

As of this commit, some important features are missing from the shell
(including any form of input). Also, the main xi-win app is not yet
using the shell.