Graph mutation
Allow mutation of the graph from inside listener contexts, not just
during building. The `UiInner` struct has been renamed `Ui`.

This patch adds an `append_child` method, and also does a bit of
refactoring so that listeners can be added. It adds remove (suitable
for moving widgets in the tree) and full deletion. When widgets are
deleted, they're fully cleaned up, with listeners removed and the id
available for resuse.

There's also an example which exercises the new APIs.