Refactor for druid structure
Lots of changes to accomodate the structure needed for xi-win to be
implemented on top of druid

Most notable changes:
`AppDispatcher` handles notifications from core, and also forwards menu
events to views.
`AppState` keeps track of the currently focused views, and is intended
as a reference for accessing views by core view_id values.
`EditView` has a weak reference to Core, allowing it to send cmds
directly.

It seems that updates caused by `EditView` commands aren't
being received properly, as the core does not respond to them.

Missing: File open, file save, file save as. Otherwise the functionality
should be the same.